New Dashboard design and implementation
There is a new dashboard design which is to include pending bookings.
The inclusion of pending bookings has resulted in another overhaul of
the data fetching logic for the dashboad.

I have started to implement the changes and the pending bookings have
been implemented correctly with its own separate data fetching logic,
however for the other components have yet to be implemented.

This changes will make use to the deprecation approach. The old logic
will stay in place until a new and better logic is implemented and
working. There will have to be changes in the database as well as soon
as the new logic is working. (Removal of the Dashboard_Overview view and
endpoint) basically returning the dashboard to its original state lmao.