Admin Side: Manage Cars | Display Car Status
Car status now displays and can be edited by those admins who are authenticated
Database has also been updated to cater such changes.
Will have to work on the auto update of the display for changes tho.

Also will have to tell Customer side devs to update their code so that whenever car's status is "In Maintenance" the customers won't be able to select the car that is under maintenance.