Admin side: Manage Bookings | Walk-In Bookings - 70% complete
Forms are complete.
Submission logic and Calculation Logic are sound.

Hard coded the booking status id to each form.
Creating a booking from customer side will set the id to 1
Creating a booking from admin side will have conditional setting. If booking is set within the same hour of pick up time then booking will be set to 3 or ongoing, otherwise it is set to 2 or confirmed.
Cash or Cashless choice payment method added for admin side.

Just need to polish this and it should be done

Remind everyone to polish the redirection after a booking has been made.